+++ marshmallow/fields.py
@@ -1106,7 +1106,7 @@ class Email(ValidatedField, String):
 class Method(Field):
     """A field that takes the value returned by a `Schema` method.
 
-    :param str method_name: The name of the Schema method from which
+    :param str serialize: The name of the Schema method from which
         to retrieve the value. The method must take an argument ``obj``
         (in addition to self) that is the object to be serialized.
     :param str deserialize: Optional name of the Schema method for deserializing
@@ -1118,15 +1118,13 @@ class Method(Field):
     .. versionchanged:: 2.3.0
         Deprecated ``method_name`` parameter in favor of ``serialize`` and allow
         ``serialize`` to not be passed at all.
+    .. versionchanged:: 3.0.0
+        Removed ``method_name`` parameter.
     """
     _CHECK_ATTRIBUTE = False
 
-    def __init__(self, serialize=None, deserialize=None, method_name=None, **kwargs):
-        if method_name is not None:
-            warnings.warn('"method_name" argument of fields.Method is deprecated. '
-                          'Use the "serialize" argument instead.', DeprecationWarning)
-
-        self.serialize_method_name = self.method_name = serialize or method_name
+    def __init__(self, serialize=None, deserialize=None, **kwargs):
+        self.serialize_method_name = serialize
         self.deserialize_method_name = deserialize
         super(Method, self).__init__(**kwargs)
 
@@ -1170,21 +1168,17 @@ class Function(Field):
         which is a dictionary of context variables passed to the deserializer.
         If no callable is provided then ```value``` will be passed through
         unchanged.
-    :param callable func: This argument is to be deprecated. It exists for
-        backwards compatiblity. Use serialize instead.
 
     .. versionchanged:: 2.3.0
         Deprecated ``func`` parameter in favor of ``serialize``.
+    .. versionchanged:: 3.0.0
+        Removed ``func`` parameter.
     """
     _CHECK_ATTRIBUTE = False
 
     def __init__(self, serialize=None, deserialize=None, func=None, **kwargs):
-        if func:
-            warnings.warn('"func" argument of fields.Function is deprecated. '
-                          'Use the "serialize" argument instead.', DeprecationWarning)
-            serialize = func
         super(Function, self).__init__(**kwargs)
-        self.serialize_func = self.func = serialize and utils.callable_or_raise(serialize)
+        self.serialize_func = serialize and utils.callable_or_raise(serialize)
         self.deserialize_func = deserialize and utils.callable_or_raise(deserialize)
 
     def _serialize(self, value, attr, obj):